requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==4.1.0
cachetools==5.3.1
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.3.2
cron-descriptor==1.4.0
cryptography==41.0.1
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.3
django-allauth==0.54.0
django-braces==1.15.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-compat==1.0.15
django-oauth-toolkit==2.3.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.6
google-api-core==2.11.1
google-api-python-client==2.93.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
httplib2==0.22.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.5.0
kombu==5.3.1
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
pluggy==1.3.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.1.0
pytest==7.4.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
stripe==5.4.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
